replay viewer: add 1/2x playback speed; Space pauses in the league shell
The replay transport gains a 0.5x speed: command '5' selects a replay-only
half speed (ReplayHalfSpeedIndex) that spends one sim tick every other frame,
the chrome grows a 0.5x chip, and the frame JSON's sp field carries 0.5 while
it is active (buildStateJson's speed is a float now). The league shell forwards
Space down its postMessage command channel so play/pause works while the shell
chrome holds focus — keydown never crosses the iframe, so the board page's own
Space binding cannot fire there.

chrome_common.js now reads tandem's own TANDEM_WIRE global (the inherited
CTF_WIRE lookup never resolved here, so the engine's PlaybackSpeeds table could
not reach the speed chips) and its speed->command map gains 0.5 -> '5'. The
viewer test re-pins the patched chrome bytes and names the three lines of the
patch, so a future re-copy of the pristine starter fails on WHAT is missing
rather than on an opaque digest.

Validation: nim c src/tandem.nim and src/tandem_player.nim; nim check
-d:emscripten replay-viewer/tandem_replay.nim; all 15 tests/*.nim green in
debug and -d:release; tools/ci/viewer_shell_check.cjs green on a locally
assembled bundle; tools/wasm_replay_smoke.cjs advanced 240 frames with no hash
mismatch against a locally built wasm module.

+++ src/tandem/broadcast.nim
@@ -264,7 +264,7 @@ proc buildStateJson*(
   sim: SimServer,
   events: JsonNode,
   playing: bool,
-  speed: int,
+  speed: float,
   maxTick: int,
   looping: bool,
   transportEnabled: bool,
